Edits 1
Did you have these in a particular order on purpose? I went ahead and moved the order around because it makes the sentence structure better, but let me know if it should be changed. Also, is there another way we can say "still a thing?" Something like "why security still matters today" or "why security is important for developers" or similar?